Eliom

Eliom is a framework for building client/server Web and mobile applications in OCaml.

It can be used both as a traditional server-side Web framework or to implement complex client-server applications.

Eliom transforms OCaml into a multi-tier language, making it possible to implement both the server and client parts of a Web and mobile app as a single program.

This simplifies a lot the communication between server and client. Applications can run on any Web browser or mobile device (iOS, Android), saving from the need to develop one version for each platform.

Eliom has support for reactive pages (generated on server or client), advanced session mechanism, server to client communication, continuation based Web programming, etc.

Eliom is part of the Ocsigen project.

Installation instructions

opam install eliom

Getting started

Defining a service on path /foo, taking any GET parameters:

let myservice =
  Service.create
    ~path:(Service.Path ["foo"])
    ~meth:(Service.Get (Parameter.any))
    ()

let () =
  Registration.Html.register ~service:myservice
    (fun get_params () ->
      Lwt.return
         Content.Html.F.(html (head (title (txt "")))
                                    (body [h1 [txt "Hello"]])))

Inserting a link towards that service, with parameters:

Content.Html.D.a ~service:myservice [txt "Home"] [("param1", "v1"); ("param2", "v2")]

Event handlers are written in OCaml:

div ~a:[a_onclick [%client (fun ev -> ... )]] [ ... ]

The client-side and server sides are written as a single program:

let%server a = ... (* code for the server part of the application *)

let%client b = ... (* code for the client part of the application *)

let%shared c = ... (* code that will be included in both parts *)

Using a server-side value in client-side code:

let%server a = ...

let%client f () =
  print_endline ~%a ; (* print in browser console *)
  ...

Calling a server function from the client program:

let%rpc f (x : int) : string Lwt.t = ... (* server-side code *)

let%client () =
  let%lwt r = f 4 in
  ...

Saving session data on the server using Eliom references:

let%server r = Reference.eref ~scope:Common.default_session_scope 0

let%server f () =
  let%lwt v = Reference.get r in
  Reference.set r (v + 1);
  ...

Where scope can be:

Common.default_session_scope (different value for each browser),

Common.default_process_scope (different value for each tab),

Common.default_group_scope (different value for each user),

Common.site_scope (value for the whole site),

Common.global_scope (global value for the whole server). Eliom references are persistant if you add optional parameter ~persistent to function Reference.eref.

Learning Eliom

More documentation here.

Write your first Web and mobile application with Eliom using Ocsigen Start.